[CWS] use 64bits cookie

+++ pkg/security/secl/model/model_unix.go
@@ -304,7 +304,7 @@ type Process struct {
 
 	CreatedAt uint64 `field:"created_at,handler:ResolveProcessCreatedAt"` // SECLDoc[created_at] Definition:`Timestamp of the creation of the process`
 
-	Cookie uint32 `field:"-"`
+	Cookie uint64 `field:"-"`
 	PPid   uint32 `field:"ppid"` // SECLDoc[ppid] Definition:`Parent process ID`
 
 	// credentials_t section of pid_cache_t
@@ -725,7 +725,7 @@ type SpliceEvent struct {
 type CgroupTracingEvent struct {
 	ContainerContext ContainerContext
 	Config           ActivityDumpLoadConfig
-	ConfigCookie     uint32
+	ConfigCookie     uint64
 }
 
 // ActivityDumpLoadConfig represents the load configuration of an activity dump
